More ECHECK flavors
Summary:
Provide the ECHECK_xx flavors (e.g. ECHECK_EQ).

This was a little more complicated than the basic ECHECK(),
and had to copy and modify some code from glog/logging.h.
But tried to keep that to a minimum.

It's not an exact match to what is in glog, so there may be future work.

Test Plan: Corresponding tests added.

Reviewers: zasgar, yzhao, #engineering

Reviewed By: yzhao, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D1794

GitOrigin-RevId: b305e40fd1052d6509a0dd6453764e818a3f2d10