Bug-fix follow-up
Summary:
Might as well use timestamp-like numbers in the test,
just in case there will be similar bugs like those discovered in D1801.

This is out of extreme paranoia, and probably won't provide much additional coverage. But the cost is also negligable, so seems worth it.

Test Plan: This is the test.

Reviewers: yzhao, zasgar, #engineering

Reviewed By: zasgar, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D1803

GitOrigin-RevId: 953d772e60282b3a55f00212ae5060a6a0719777