Test case showing parse difference between byte vs string
Summary:
Test that shows difference between a message field being declared as byte vs string.

This is a real-life case I ran into while performing analysis.

It's a complex message, so it's not clear which part of the message is treated differently,
but something is different.

Test Plan: The whole thing is a test.

Reviewers: yzhao, zasgar, #engineering

Reviewed By: zasgar, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D1781

GitOrigin-RevId: 3041bad0851d1e77619a92a50c0d5102bf965121