Record actual size.
Summary:
This is for visibility into the issue. We observed that crash from multiple response messages for
gRPC stream, ensues a truncated message.

The code should handle the obvious parse failure through resyncing the stream.
But there can be cases that the interaction with the followup processing is not well understood yet.
Investigation should be left for future work.

Test Plan: Jenkins

Reviewers: oazizi, #engineering

Reviewed By: oazizi, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D1819

GitOrigin-RevId: 12f8f3fd791039823ada2dc733822c7416d1651c