Disable connection trackers if EOS condition is met
Summary:
This diff brings together a few recent pieces.

If a connection is detected to be at end-of-stream (EOS), then it diables the tracker.

Practically, this is useful for HTTP upgrade requests, where the connection is changing protocols.

It is left as future work to support cases where the protocol changes to something else that we know how to handle (e.g. HTTP2)

Test Plan: Basic test case in connection_tracker_test.cc

Reviewers: yzhao, #engineering

Reviewed By: yzhao, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D1820

GitOrigin-RevId: 1edf7d50d5da5ddd7f263239bd08a60f7352726a