UI: Fix vizier router paths
Summary:
In the Nginx fix, I updated the UI to handle paths in a way that worked when running the UI locally... which turned out not to work in the gke environment because the "matchPath" is different from running in the local env.
Updated the routes so that they work for both local and gke environments.

Test Plan: jest test

Reviewers: zasgar, #engineering

Reviewed By: zasgar, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D1829

GitOrigin-RevId: 30fea7cfb8ef589d59ff26cb00c20f782111409d