String data not marshalled to JSON correctly
Summary:
in https://phab.corp.pixielabs.ai/rPLM660d2d1435922372fdd7d444f31bcbec6781912c, we updated the stringdata in our table schema to use bytes instead of string.
this is not marshalled correctly to JSON, so all of our string data was unreadable in the UI.
added a customtype to string data so that it marshals the way we want.

Test Plan: unit test and also started up minikube to make sure i didnt break anything

Reviewers: zasgar, oazizi, #engineering

Reviewed By: oazizi, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D1848

GitOrigin-RevId: 4104955e36521428739dfe78c42546a9b39e9c01