Stirling: Change behavior on No content-length
Summary:
If we see an HTTP message with:
1) No content-length or transfer-encoding, and
2) Is not a request where we can assume no body (e.g. GET)
3) Is not a response where we can assume no body (e.g. special status codes)

Then the end of the message is supposed to be signaled by the connection close.

For simplicity, we just output whatever contents we have captured so far, and print a warning.

The assumption is that a connection close has already occured, or will be occuring soon.

Test Plan: Some tests added to http_parse_test.cc

Reviewers: yzhao, #engineering

Reviewed By: yzhao, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D1865

GitOrigin-RevId: 48de5ef22518ad82f5b11d4de12d3c78520f2843