set up yaml for log forwarding pipeline
Summary:
We need to have a cloud connector service for the vizier pods to talk to,
so this diff adds that. We also need the various services in vizier to know what is the
address of that cloud connector service, so that is added here as well. finally,
the pods will need to send their pod name along with logs so that we get context on
where the logs are coming from, and that is done via the use of an environment variable.

Test Plan: n/a

Reviewers: michelle, zasgar, #engineering

Reviewed By: zasgar, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D1875

GitOrigin-RevId: 60614294daa15bd38755fdf7b338737ae2aa1cda