PL-894 Encrypt VizierCluster JWTSigningKey in postgres
Summary:
Uses pgcrypto to encrypt the JWT signing key.
When adding the signing key to the database, we add a salt to the front of the jwtkey.
The salt + jwtkey is then encrypted by a key that should be passed in using pl-app-secrets (this still needs to be added to the new version of load_secrets)

When getting the key out of the database, postgres decrypts the value using the key. The jwtkey is the value with the salt removed.

Test Plan: added tests

Reviewers: zasgar, #engineering

Reviewed By: zasgar, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D1876

GitOrigin-RevId: b784467c5b38622a75320dc442f066eef2355af7