Admin CLI: Add download-ca cmd
Summary: users need a way to get Vizier's CA into a file. Once they have the CA, they can add it as a trusted CA in their browser

Test Plan: ran it. tried loading the cert, but still need to figure out how to fix the cert to allow the IP

Reviewers: zasgar, #engineering

Reviewed By: zasgar, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D1879

GitOrigin-RevId: 0e8f826976819929d88897332b2466085e0ac89d