Fix metadata mockgen
Summary:
mockgen was broken for the metadata controllers because of the test_proto.go file, which wasnt a *_test.go file, but was in the test_controllers package.
moved it into a separate directory instead, because that seemed to make more sense than renaming it to something like test_proto_test.go...
unfortunately, this means i had to go and add comments for each exported PB variable.

Test Plan: ran go generate, verify that tests still pass

Reviewers: zasgar, oazizi, #engineering

Reviewed By: zasgar, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D1898

GitOrigin-RevId: 2d48a5b5e8b9ff4a4e23929cf7175e7a6cdd29ae