Create cluster if does not exist
Summary: Before we were trying to create it after the user creates the org for the first time, but that can be error prone if that request fails. instead, if they've made it to the vizier page and no cluster exists, just create it for them.

Test Plan: tried it manually, jest test

Reviewers: zasgar, #engineering

Reviewed By: zasgar, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D1926

GitOrigin-RevId: 8684fe2b53e0e2fc124a2c977d85cad1cf8591ec