Remap query-created tables to write to tables related to query id rather than table name.
Summary: Needed to fix an issue where tables written  to the same name by queries that didn't share a relation were continually failing. We solved this before, but changing table_store caused this to go into regression, so found a different solution. As we don't wanna support saveable tables in the way it was previously setup, this is a bit closer to what we'll end up using when we do support saving tables.

Test Plan: Updated carnot test.

Reviewers: michelle, zasgar, oazizi, #engineering

Reviewed By: oazizi, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D1929

GitOrigin-RevId: 724fd916b1cef7fccadc39f0bcccdefc0710eb5e