Let StitchFrames() return ParseState
Summary:
So that when it detects incorrect frames sequence, it signals caller to reset the streams.
This works in the same fashion as the case of duplicate responses on the stream.

In the process, also simplified the code to remove progress tracking, as we only need to record
if the parser is currently in an opened header block. So use a bool to replace the enum.

Bundled with random cleanups does not change logic.

Test Plan: Jenkins

Reviewers: oazizi, #engineering

Reviewed By: oazizi, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D1906

GitOrigin-RevId: c19cb66370ab8d43a6076b1bbc72b4fd8629b9c2