Specify cloud addr in Pixie CLI deploy cmd
Summary:
We need to be able to specify the address of Pixie cloud that the vizier cloud connector should be talking to.
before, I was hardcoding the cloud address based on the environment (staging, prod, testing, etc), but that's not right.
Instead, add a flag to the Pixie CLI which creates a new config map.

Test Plan: used deploy to load the config map, launched vizier to make sure it was using the correct address in the config map

Reviewers: zasgar, #engineering

Reviewed By: zasgar, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D1955

GitOrigin-RevId: fd52b3c28fefd3a471174036743a83c21e14487f