CloudConn should not crash if VzProxy's External IP not available
Summary:
when Vizier starts up, it takes about a minute for the proxy's external IP to become available. as a result, the cloudconnector keeps crashing because it cant get the external IP.
this leads to a crashLoopBackOff which cna be really annoying. instead, we should still send heartbeats to VZConn, but it should detect the status as "unhealthy"

Test Plan: unit test

Reviewers: zasgar, #engineering

Reviewed By: zasgar, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D1956

GitOrigin-RevId: 610a1682e9efdf2ca03c31a58c6b39b5d6a4dfce