Stirling/BPF Optimization: filter read/write of files
Summary:
This is for optimization purposes only.
We can skip some BPF processing when we know the FD belongs to a file instead of a socket.

We detect this through a call to open() vs a call to connect()

While at it, I pulled out id into the wrappers, for consistency reasons.

Test Plan: Existing tests

Reviewers: yzhao, #engineering

Reviewed By: yzhao, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D1935

GitOrigin-RevId: 4e861bec0a2b492d1814c07a394619dc3d3c42d7