CLI: Wait until VZProxy external IP is available before exiting CLI
Summary:
After we finish deploying Vizier, it takes some time for Vizier to actually work in Pixie cloud. This is because the external IP takes some time to be assigned.
the CLI shouldnt exit until that is ready, so that the user doesn't get confused about why Pixie Cloud hasnt connected to Vizier yet.

Test Plan: ran it manually

Reviewers: zasgar, #engineering

Reviewed By: zasgar, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D1998

GitOrigin-RevId: 566c1a48fe03ede659c06439c2513ba5d7fc973d