Proc parser of FD links
Summary:
Simple function to run readlink on the /proc/<pid>/fd/<fd>.

Required to get the inode of socket FDs, when trying to determine the remote endpoint of established TCP connections.

Test Plan:
Tried adding a test, but since this function uses readlink, and bazel creates an extra layer of symlinks, the test is broken.
I have tested it manually though.

I've left the broken test in for reference, but in a disabled state.

I also tried resolving the symlink in the test itself, but bazel creates individual symlinks for each file rather than a symlink for the entire testdata directory. Otherwise, that might have worked.

Reviewers: yzhao, zasgar, #engineering

Reviewed By: yzhao, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D2005

GitOrigin-RevId: 8e1824888fbea5d941feb9d399b0ab04fa3e9992