Bug-fix in readlink argument (reserve vs resize)
Summary:
Changed from resize() to reserve() as part of the last review,
but turns out the original is the correct way.

I was getting weird behavior and root caused to this change.

I agree with this, btw:
https://herbsutter.com/2019/09/10/qa-does-stringdata-return-a-pointer-valid-for-size-elements-or-capacity-elements/

Semantically, resize() is the right thing to call before passing the buffer to the function.

Test Plan: Manually tested.

Reviewers: yzhao, #engineering

Reviewed By: yzhao, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D2010

GitOrigin-RevId: b5d2a509cb8d3f32585a4c479271bf4cc3b57696