Fix issue with registry info memory.
Summary:
Had a unique ptr within a function that got deallocated earlier than the lifecycle of the memory it used. Re shuffled args.
Also snuck in a little DCHECK removal because the DCHECK wasn't necessary -the rest of the code continued to function if the DCHECK fails.

Test Plan: Add new test to make sure the udf stuff doesn't regress

Reviewers: michelle, zasgar, #engineering

Reviewed By: michelle, #engineering

Subscribers: yzhao

Differential Revision: https://phab.corp.pixielabs.ai/D2047

GitOrigin-RevId: 79a612f9c580bf4f54e8348acec7a08a820aa71e