Add dnsmgr service skeleton
Summary:
We need a service in cloud which is responsible for assigning DNS addresses to clusters, and talking to google cloud DNS to update records.
This is just setting up the skeleton for the service. the service currently does nothing.
we probably also need to add a database to track the available DNS addresses and which clusters are assigned to them.

Test Plan: ran in skaffold and made sure the service starts up

Reviewers: zasgar, #engineering

Reviewed By: zasgar, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D2072

GitOrigin-RevId: 144dfa159aebdb45f03bcf0ce9589232ebee2dc2