SocketResolver: to identify remote endpoint when missing
Summary:
If we miss the accept/connect, then we don't know the endpoint that a traced message was talking to.

This diff adds a SocketResolver class, and expands the ConnectionTracker to use the SocketResolver as it tries to infer the remote endpoint.

SocketResolver is made complicated by the fact that we can only query a socket by its FD, and FDs can be reused.
Moreover there is no known way to query Linux for the creation time of a socket.

This mean we have to make sure we are assigning the endpoint to the right connection.

This is done by monitoring the FD over two or more samples. If a traced event on the FD occurs between the time period in question, then we can assume we have the endpoint information for the correct connection.

Test Plan: Tests added for SocketResolver.

Reviewers: yzhao, zasgar, #engineering

Reviewed By: yzhao, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D2062

GitOrigin-RevId: 8ca2adf7c695be7bf2c79e1236fc1a9dcfc86b86