Show DNS loading message when vz is not connected
Summary: It takes some time for us to set up the DNS for Vizier. we should should the users a loading message while this is happening.

Test Plan: unit test. was kind of able to try it locally by manually updating database records, but still needs to be checked in a working env

Reviewers: zasgar, #engineering

Reviewed By: zasgar, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D2109

GitOrigin-RevId: b256091f54456d557c10930e590298a83616319d