Fix issue with Join where it would crash without any feedback to the query console.
Summary:
Join would crash if an output column that doesn't exist would be present in a query, instead of returning an error. This was caused by a missing check on whether column data types had been resolved in the Join query. There was a possibility that this bug could cascade issues elsewhere so it's good this
was addressed.

Test Plan: added new test to make sure this issue doesn't happen again

Reviewers: michelle, zasgar, #engineering

Reviewed By: michelle, #engineering

Subscribers: yzhao

Differential Revision: https://phab.corp.pixielabs.ai/D2145

GitOrigin-RevId: f463d028823855f9c8d9f71af9f5735c9aad4b6d