BPF: new conn_event_t and close_event_t structs to optimize transfer through perf buffer
Summary:
We were previously sending conn_info_t to communicate opens and closes of connections,
but that is really an internal tracking, which has more information than needs to be transmitted.

This diff creates auxiliary structures explicitly for communication with user space.
It causes some extra copying in BPF, but saves on copies later in user space, because of the smaller structure.

Bonus: Also changed the passing of these structs to const references in socket_trace_connector and connection_tracker to minimize copying. It required moving the application of the timestamp offset.

Test Plan: Existing tests

Reviewers: yzhao, #engineering

Reviewed By: yzhao, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D2149

GitOrigin-RevId: 2617a60f8e585ee99be2391209f681a98d8bbd3a