Append port to Vizier address
Summary:
We need the port to be included in the address that we send to the UI.
Updated cloudpb.VizierHeartbeat to include a port field, which is then added to the end of the address when storing it in the db.
Another way we could have done this was to not add a new port field to the pb, but append it to the address field in cloudpb.VizierHeartbeat.
However, that would mean that vzmgr would have to parse out the port, since it needs to be able to get just the IP address to send to DNSMgr. This seemed a little messier to me.

Test Plan: updated unit tests

Reviewers: zasgar, oazizi, #engineering

Reviewed By: zasgar, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D2181

GitOrigin-RevId: 4c32a812cf966dbc1be6ae796b2f60cfa5d1a865