K8s plumbing for running the headerless agent
Summary:
This is used for individual use by engineers to run agent for a long time. Currently primarily for
verifying the stability of table store and stirling components.

Tweaked organization of ymal files.
The shell script is based on src/stirling/yaml/run_on_gke.sh, with simplifications:
* Delete namespace entirely.
* Omit some error check on argument.

Test Plan:
Manual compare the output of kustomize build k8s/vizier/dev on mainline and with this diff.
The only diff are the reordered certs specs.

$ diff a b
517a518,519
>         - mountPath: /certs
>           name: certs
524,525d525
<         - mountPath: /certs
<           name: certs
535a536,538
>       - name: certs
>         secret:
>           secretName: service-tls-certs
544,546d546
<       - name: certs
<         secret:
<           secretName: service-tls-certs

Reviewers: oazizi, zasgar, #engineering

Reviewed By: oazizi, zasgar, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D2195

GitOrigin-RevId: 830aa5aea37939a56a3262d688030f5ddd3af6ec