Stop localhost errors
Summary:
we were logging out tons of network errors when trying to connect to localhost.
instead, as soon as we successfully connect to vizier, we should not try to connecting to localhost anymore.
if vizier's request fails, we should continue trying localhost until something works.

Test Plan: tried it manually

Reviewers: zasgar, #engineering

Reviewed By: zasgar, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D2234

GitOrigin-RevId: 71287162a35a232fdace639ef6e8abfb315dc5c0