Implement LogWriter for vizier
Summary:
This logwriter implements io.Writer so it can be passed into logrus as a
destination for vizier logs. It batches them up and sends them to the
cloud connector, which will then forward them to pixie cloud.

It handles cases where the main program crashes, and can support both a
maximum time interval to hold onto messages as well as a maximum queue
size.

Depends on D1874.

Test Plan: added

Reviewers: zasgar, michelle, #engineering

Reviewed By: zasgar, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D2194

GitOrigin-RevId: da3d166703c2f70879c1080f7592ea3fcb358187