MySQL Stitcher: Include full list of commands
Summary:
There should be a stitcher for each possible type of command. This add the full list of commands,
and include a stitcher for commands with a single-packet response.

New commands with more complex responses are left as future work.

Test Plan: Added a stitcher test case for the basic request-response, using COM_PING as an example.

Reviewers: yzhao, #engineering

Reviewed By: yzhao, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D2242

GitOrigin-RevId: 376f7258aa8d53dfaa5c15b370836e9f8a9c0799