Remove old compiler export
Summary: Compiler export was the old way of compiling queries in the query broker. Removing as it creates unnecessary CGO dependencies on anything that imports the compiler in GO. Note that this deleted functionality is now in the logical planner which currently contains CGO.

Test Plan: removed and didn't break anything - in fact fixed a docker image that depended upon this unintentionally.

Reviewers: zasgar, michelle, nserrino, #engineering

Reviewed By: michelle, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D2255

GitOrigin-RevId: d2d8e1c6d6cc45eebc500da701b55a1a473a4cef