Fix flaky leader election test
Summary:
The leader election test is flaky in master. The problem is that we are using an embedded etcdserver in the test, which can sometimes take longer to respond (up to 900 ms)
however, the Campaign function in the leader election module has a timeout, since we dont want the metadata service to just hang and wait for it to become leader. In production, this timeout is okay because we're executing the campaign in a loop in RunElection. however, in the test, we want to test just the Campaign function and dont want it to timeout. added a function so that we can specify a longer timeout.

Test Plan: ran the test with --runs_per_test=400

Reviewers: zasgar, #engineering

Reviewed By: zasgar, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D2259

GitOrigin-RevId: ef722e8d7039ab62edd0481c40ecb43e62ab7bca