Add instructions on adding and updating submodule
Summary:
This is not detailed enough. Later when I have time, I'll add more details.

So there was a `pixie-head` added to protobuf fork, which is different from the
original `pixielabs` designated to be the tracking branch we use.

I think we should avoid adding more than one new branch to the submodule.
So I removed the `pixielabs` branch and created this diff.

Test Plan: Not needed

Reviewers: oazizi, zasgar, #engineering

Reviewed By: zasgar, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D2252

GitOrigin-RevId: 2dc1e9022af270560c2022d46ed86d33579cbffb