Fix segfault caused by data initialization order
Summary:
Ran into an issue in another build where LOG was causing an exception.

Appears as though the logging environment was not set-up.

Caused by a call to LOG (from TestEnvironment::PathToTestDataFile()) as part of a static class member initialization, which would occur before the logging environment is up.

Test Plan: Jenkins

Reviewers: zasgar, yzhao, #engineering

Reviewed By: zasgar, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D2283

GitOrigin-RevId: e436740065f2885fa6d7f8c0c1b4174f26315071