Fix rendering of floating-point rounding errors after dividing 0 by a number to display 0 instead of NaNs
Summary:
dividing 0 by a non-0 yielded nans because of floating point errors. Fixed this by clipping any small float value to 0.

nans: {F27428}

to no nans
{F27427}

This is related to this open issue for Numeral.js
https://github.com/adamwdraper/Numeral-js/issues/512

On a related note the author hasn't touched the repo in 2 years now.

Test Plan: tried it in the UI

Reviewers: michelle, zasgar, #engineering

Reviewed By: michelle, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D2292

GitOrigin-RevId: 4db338391aea4179047e82b2355b172327ee76cc