Fix the bad Nan rendering commit
Summary: I messed up the IsNan commit. Turns out that anything that's not a float or an int evals IsNan to true - meaning anytime you pass in a Numeral object it is NaN, although semantically it is not. Fixed now to check if the output of numeral is NaN instead.

Test Plan: tested with the golden queries - no NaNs and more interesting numbers than just 0.

Reviewers: michelle, zasgar, #engineering

Reviewed By: michelle, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D2311

GitOrigin-RevId: 05fe2fd60891d685777d5318b649c10ea3ff004d