MySQL: Revamp stitcher/handler.
Summary:
To make the code easier to maintain, the stitcher now takes one req_packet,
and collects a "view" of attached resp_packets into a bundle, and sends these for processing.

The processing functions are no longer responsible for popping resp_packets, which was previously haphazard.
Instead the resp_packets are used in a read-only way to collect the relvant response data.

At the end of the stitcher, unless the result was kNeedsMoreData as the last bundle, the the stitcher pops off the req_packet and attached resp_packets,
and moves on to the next bundle.

Test Plan: Updated existing tests

Reviewers: yzhao, #engineering

Reviewed By: yzhao, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D2318

GitOrigin-RevId: 2b60328502e1d052a9e91363188ee72cc3302e44