Update DeployInstructions to use CliArtifact gql endpoint
Summary:
Before we hardcoded the download link for the CLI binary/SHA256.
Now that we have a gql endpoint which returns the download link for the latest CLIArtifact/SHA256, we should use that instead.
This way, even if we don't update cloud, but update the CLI, the UI will always show the link to the latest CLI binary.

in the future we need to clean up the deploy instructions to allow the user to select whether they want to download the linux or CLI binary.

Test Plan: tried it manually, updated and added unit tests

Reviewers: zasgar, malthus, #engineering

Reviewed By: zasgar, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D2350

GitOrigin-RevId: ff0d2b9f98fe7861e403660a808059e5b7ea880b