Add script to get # of commits between tags
Summary:
We need a script to figure out the number of commits between a tag and the previous tag.
This script finds all the tags for a release type (for example, release/cli vs release/vizier) in sorted order. it then finds the previous tag in the list of tags and gets the count of commits.

Test Plan: tried it manually. ./get_commits_since_release "release/cli-v0.1.4.12-a"

Reviewers: zasgar, #engineering

Reviewed By: zasgar, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D2351

GitOrigin-RevId: 81948a9cec56791ec9ebff2490d8b41accd6e3aa