Add script for creating new release tag
Summary:
added a script that generates a new release tag using the provided input flags.
if the user is creating an external release, they can specify -m or -n to bump the major/minor version. if neither -m or -n are specified, the patch version is bumped.
otherwise, if it is an internal release, we count the number of commits since the last release tag and use that as the -pre.<commit-num> part of the tag.

the changelog is given through a pipe
echo 'this is a change' | ./create_new_release cli -p -r -m
The above command creates a new release tag with the major version bumped and 'this is a change' as the tag message. the tag is also pushed.

Test Plan: tried it manually

Reviewers: zasgar, #engineering

Reviewed By: zasgar, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D2363

GitOrigin-RevId: 7ff27bbf86a620c1e8aeaa21acdd64452abe77dc