Stirling: Reorganize protocol parsers into own directories
Summary:
In this diff, protocol parsers are each moved to their own sub-directory.
 - common: any shared classes/structs/interfaces.
 - http: all HTTP parsing related code
 - http2: all HTTP2 and GRPC parsing related code
 - mysql: all MySQL parsing related code.

Note: the structure of files within the sub-directories is not currently standardized, but that is left as a separate task.

Test Plan: Existing tests. Had to disable the event_parser_test, which I need to re-work before landing.

Reviewers: yzhao, #engineering

Reviewed By: yzhao, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D2357

GitOrigin-RevId: 0ce7d7bf4938ef5919dec5d145007cd22928fae4