Add elfio repository
Summary:
This is a package that can be used to get the symbols tables out of binaries.
Useful for upcoming experimental code, where we want to search for GRPC symbols to trace.

See https://phab.corp.pixielabs.ai/D2395 for where it will be used.

Test Plan: Manually tested that repository works.

Reviewers: yzhao, zasgar, #engineering

Reviewed By: zasgar, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D2394

GitOrigin-RevId: a0e0a40878ef1d84fe987c8e1bc19b082c827e4d