Update Pixie deploy to use Vizier YAMLs artifact
Summary:
Now that we have the Vizer YAML artifacts, we no longer need to use the YAMLs packaged with our CLI.
Instead, we should query the artifact tracker service for the download link, get the contents out of the tar, then deploy the yamls.

if no version is specified, we can assume the user wants to deploy the latest version of Vizier.

this code should be updated to use the download util whenever its ready.

we also use the go generated YAMLs in Pixie's delete cmd, which will be addressed in a followup diff.

Test Plan: tried it manually with and without specifying --use_version

Reviewers: zasgar, #engineering

Reviewed By: zasgar, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D2390

GitOrigin-RevId: 9d2262bf1ca8bdfd3a47a10d73080b09ef8c3265