GRPC prober: scan for a symbol across active binaries
Summary:
This version can scan all existing binaries to search for a symbol pattern of interest.
This is important for GRPC probing, where code is generated, and we are looking for
a templated function.

Where the relevant function is found, a probe is attached.

Test Plan: Experimental. Loosely manually tested.

Reviewers: yzhao, zasgar, #engineering

Reviewed By: yzhao, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D2395

GitOrigin-RevId: 5fc622912ce0219ef6985d49b9e9c9a9457f0bb9