Add APIs for attaching uprobes to BCCWrapper
Summary:
Wrapping around the BPF::attach_uprobe() following the existing pattern for the kprobes.

I'll defer the testing. As soon we'll have the uprobes added to socket_tracer.c, and one of the
bpf test files. With that we'd be able to test everything on top of the existing test fixture.

Test Plan: Jenkins

Reviewers: oazizi, #engineering

Reviewed By: oazizi, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D2408

GitOrigin-RevId: 32a522f866801ccd99c7f50d76d621007059813e