Refactor the company login component to handle errors better.
Summary: The issue is that the submit button is disabled if there is an error. This isn't great since any network error would require refreshing the page to re-submit. Added logic to disable the button and input field if the request is in flight, and refactor the code a bit.

Test Plan: Tested the UI manually, updated the unit tests for the component.

Reviewers: michelle, #engineering

Reviewed By: michelle, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D2425

GitOrigin-RevId: 1b3b57b12f3102c1bec242d6ef43761e5c5a2868