Add begin_timestamp_ns to socket_data_event_t::attr_t
Summary:
Represents when the entry probe is triggered.

I opt not to change timestamp_ns to end_timestamp_ns (or something along the line).

It should be straightforward to derive the semantic when is in contrast to begin_timestamp_ns.

The follow up would be to use being_timestamp_ns to produce a more accurate timespan for a higher
level message.

Test Plan: Jenkins

Reviewers: oazizi, #engineering

Reviewed By: oazizi, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D2432

GitOrigin-RevId: 413a2d84e25efba4b9088eeaf949162fc5a18d8e