Create IP->Hostname Map in etcd
Summary:
When sending updates to agents, we need to send it pods that are/were running on the agent's node.
the current logic for this is buggy for pods that used to run on the node. we need a map that maps ip to hostname so that
we can find the right hostname for deleted pods.
this diff is just for creating that map.
the actual fix will be in a followup diff

Test Plan: unit tests, ran in skaffold and checked that the IP address for minikube was written correctly in etcd

Reviewers: zasgar, #engineering

Reviewed By: zasgar, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D2441

GitOrigin-RevId: ffe88e8e9aa8ff982ae407e818ced0f1c9ee8d02