Add gRPC API to access Vizier cluster information
Summary: This will let us fetch vizier info from non-UI interfaces.

Test Plan:
We need to figure out a scheme to share gRPC and graphql code. Going to do that next after this and will then
figure out a good way to test this. For now I tested it manually.

Reviewers: michelle, #engineering

Reviewed By: michelle, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D2444

GitOrigin-RevId: febb85e4e2e554164b8b90e26ded1e04f707f17d