Add Pixie CLI flag for deploying vizier on Cloud minikube
Summary:
the current cloudAddress we use for the cloudConnector to talk to VzConn doesn't work if cloud is running on minikube.
this is because we set dev.withpixie.dev to point to the proxy's externalIP in etc/hosts. however, we can't get to Vzconn's address from this IP.
updated the deploy and load-cluster-secrets command to take a new flag where the developer can specify the namespace of Pixie Cloud if its running on minikube.
Using this namespace, the cloudConnector can use VZConn's FQDN.

Test Plan:
deployed Vizier on minikube using cloud running on minikube.
deployed Vizier on minikube using prod cloud.

Reviewers: zasgar, #engineering

Reviewed By: zasgar, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D2478

GitOrigin-RevId: 5f974051c278959ec535510d819ca4d087abf6c0