Func refactor diff 2: Add rule to evaluate compile time functions
Summary:
Depends on D2484. This diff is part of a series of diffs that refactors the way that we
handle funcs, specifically compile time constant evaluation. This diff introduces a rule
to evaluate compile time expressions (but does not yet integrate it into the code).
This addition of this rule informs a future refactor of rules to be id-based in their input
rather than pointer-based for increased composability.

Test Plan: added test

Reviewers: philkuz, michelle, zasgar, #engineering

Reviewed By: philkuz, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D2485

GitOrigin-RevId: dc1b5d7299d02efaf5679cf76f4a590f3bc74b16