Func refactor diff 5: Refactor FuncIR Init() call to not use unused params prefix and compile_time
Summary:
Depends on D2488. This diff is part of a series of diffs that refactors the way that we
handle funcs, specifically compile time constant evaluation. This diff changes the signature of
FuncIR's Init to no longer use the unneeded params prefix (since funcs will all be under the same prefix now
from the func IR's perspective) and compile_time (since whether or not it's a compile time func will be
evaluated as part of a rule rather than a special case for RangeArgs).

Test Plan: existing

Reviewers: philkuz, michelle, zasgar, #engineering

Reviewed By: philkuz, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D2490

GitOrigin-RevId: b14835951a7e04440a274dfc3e9f31d0afab9a0e