Func refactor diff 6: Refactor ast_visitor to produce Func expressions for range agg args.
Summary:
Depends on D2490. This diff is part of a series of diffs that refactors the way that we
handle funcs, specifically compile time constant evaluation. This is where the refactor actually gets propagated
to the code that is executed in Carnot, in Range arg. This generalization paves the way for non-lambda expressions
and compile-time expression evaluation to be used in the same way in filter, map, etc.

Test Plan: removed some tests that no longer apply :P

Reviewers: philkuz, michelle, zasgar, #engineering

Reviewed By: philkuz, #engineering

Subscribers: yzhao

Differential Revision: https://phab.corp.pixielabs.ai/D2492

GitOrigin-RevId: e4e38486972e64d51fcb9cba577fd609b48b9ab0