PL-1148 Fix agents crashing because of dangling reference
Summary:
When an agent connects to the metadata service for the first time, the metadata service sends over all relevant pods and services.
Because of a fix last week, which actually only worked on minikube, the metadata service was not sending over any pods at all. This would result in a dangling reference error where the agent would crash.

Also ran into another timing issue with leader election. The metadata service would not block on its first leader campaign, so it might skip the step where it is supposed to sync with etcd because it hits that step before it becomes leader.,

Test Plan: ran on gke without any crashes, unit test

Reviewers: zasgar, #engineering

Reviewed By: zasgar, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D2507

GitOrigin-RevId: f0989ea6030aada68bde0863f2da50fecf534dea